Create custome NumberTextField which solves a bug regarding min/max
The implementations hides the goofy inputProps property for specifying
min and max. When they were set, it would only bound the numbers when
using the click increment/decrement but not typing a number in. I gave
the component a state that the rest of the application wouldn't care
about so it uses remembers the last valid number so if a non-valid
number is input, it swaps it.