Add api router to the server and a room code generator
The room code is just a bijection using the fact that adding a coprime
of 26 onto the set from 0 to 25 will looping on that numberset will
create a reordering of numbers. Do that with a different coprime on
each letter will give us a seemingly random code on every generation
while keeping the full set of 26^4 unique codes and increments before
starting over. Creates the nextRoomCode collection in mongo if it doesn't already
exist.